django_app/requirements.txt
amqp==2.5.2
billiard==3.6.1.0
Django==2.2.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
importlib-metadata==1.3.0
kombu==4.6.7
more-itertools==8.0.2
psycopg2-binary==2.8.4
pytz==2019.3
redis==3.3.11
sqlparse==0.3.0
vine==1.3.0
zipp==0.6.0
celery==4.4.0
boto3==1.10.42
django-celery-beat==1.5.0